WIP: Update Transfer Knight to include changes to owner_to_knights hash
- re: no action on second match statement.
The Err is actually the desired behavior, as that
would be the case where the id isn't found in the
list of ids. If it were found that would mean the
&to account already owns that knight. Other examples
I've seen include mutating the Vec in the Err(), but
I couldn't get this to work. I tried something like
ids.as_mut()... but this change didn't persist.
I also tried OwnerToKnights::<T>::append(&to, id);
in the lamdba but this didn't persist either.
This version is the only way I could get it to work.
I am committing it in this form so I can come back
to it later and improve it. I do believe I need
better error handling here, too, to make sure
state is rolled back in the event of an error.